+++ pytorch_lightning/profiler/__init__.py
@@ -1,5 +1,5 @@
 from pytorch_lightning.profiler.advanced import AdvancedProfiler
-from pytorch_lightning.profiler.base import AbstractProfiler, BaseProfiler, PassThroughProfiler
+from pytorch_lightning.profiler.base import AbstractProfiler, BaseProfiler, PassThroughProfiler, Profiler
 from pytorch_lightning.profiler.pytorch import PyTorchProfiler
 from pytorch_lightning.profiler.simple import SimpleProfiler
 from pytorch_lightning.profiler.xla import XLAProfiler
@@ -7,6 +7,7 @@ from pytorch_lightning.profiler.xla import XLAProfiler
 __all__ = [
     "AbstractProfiler",
     "BaseProfiler",
+    "Profiler",
     "AdvancedProfiler",
     "PassThroughProfiler",
     "PyTorchProfiler",